# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spacze AI Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "1536x1024": "1536x1024",
  "1024x1536": "1024x1536",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  scaffolding: "scaffolding",
  ready: "ready",
  error: "error",
} as const;

export interface Project {
  id: number;
  name: string;
  description: string;
  framework: string;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  path: string;
  content: string;
  language: string;
  createdAt: string;
  updatedAt: string;
}

export type ProjectWithFilesStatus =
  (typeof ProjectWithFilesStatus)[keyof typeof ProjectWithFilesStatus];

export const ProjectWithFilesStatus = {
  scaffolding: "scaffolding",
  ready: "ready",
  error: "error",
} as const;

export interface ProjectWithFiles {
  id: number;
  name: string;
  description: string;
  framework: string;
  status: ProjectWithFilesStatus;
  createdAt: string;
  updatedAt: string;
  files: ProjectFile[];
}

export interface CreateProjectBody {
  name: string;
  description: string;
  framework: string;
}

export interface UpdateProjectFileBody {
  content: string;
}

export interface GenerateProjectCodeBody {
  prompt: string;
}

export interface DebugProjectCodeBody {
  error: string;
  fileId?: number;
}

export type ProjectStatsByFramework = { [key: string]: number };

export type ProjectStatsByStatus = { [key: string]: number };

export interface ProjectStats {
  total: number;
  byFramework: ProjectStatsByFramework;
  byStatus: ProjectStatsByStatus;
  recentProjects: Project[];
}