# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spacze AI Agent API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat and code generation operations
  - name: projects
    description: AI-generated project management

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image/icon from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOpenaiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateOpenaiImageResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all AI-generated projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project scaffold via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project with files
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithFiles"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [projects]
      summary: List all files in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of project files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"

  /projects/{id}/files/{fileId}:
    put:
      operationId: updateProjectFile
      tags: [projects]
      summary: Update content of a project file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectFileBody"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

  /projects/{id}/generate:
    post:
      operationId: generateProjectCode
      tags: [projects]
      summary: Generate or regenerate code for a project via streaming AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateProjectCodeBody"
      responses:
        "200":
          description: SSE stream of code generation progress
          content:
            text/event-stream: {}

  /projects/{id}/debug:
    post:
      operationId: debugProjectCode
      tags: [projects]
      summary: Debug code errors using AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebugProjectCodeBody"
      responses:
        "200":
          description: SSE stream of debug analysis and fixes
          content:
            text/event-stream: {}

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project statistics and overview
      responses:
        "200":
          description: Project stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    GenerateOpenaiImageBody:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
      required:
        - prompt

    GenerateOpenaiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        framework:
          type: string
        status:
          type: string
          enum: [scaffolding, ready, error]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - framework
        - status
        - createdAt
        - updatedAt

    ProjectFile:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        path:
          type: string
        content:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - path
        - content
        - language
        - createdAt
        - updatedAt

    ProjectWithFiles:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        framework:
          type: string
        status:
          type: string
          enum: [scaffolding, ready, error]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        files:
          type: array
          items:
            $ref: "#/components/schemas/ProjectFile"
      required:
        - id
        - name
        - description
        - framework
        - status
        - createdAt
        - updatedAt
        - files

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        framework:
          type: string
      required:
        - name
        - description
        - framework

    UpdateProjectFileBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    GenerateProjectCodeBody:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt

    DebugProjectCodeBody:
      type: object
      properties:
        error:
          type: string
        fileId:
          type: integer
      required:
        - error

    ProjectStats:
      type: object
      properties:
        total:
          type: integer
        byFramework:
          type: object
          additionalProperties:
            type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer
        recentProjects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
      required:
        - total
        - byFramework
        - byStatus
        - recentProjects